- Added a Batch method to test if an object can be added in the batch based on its memory usage and number of textures - Define for max of texture per batch - Resolved some warnings - Added GPU markers for batch rendering - TODO: Need to modify Mesh class to not be created on GPU if they are used in Batching ( fullscreen quads are not used in batch so we have to create) - TODO: Create a default material

+++ VULKAN/DescriptorsUtils.cpp
@@ -110,7 +110,7 @@ void DescriptorPool::Construct(const std::vector<DescriptorSetLayout>& layouts, 
 			poolSize.push_back(pSize);
 		}
 
-	m_remainingSets = maxSets * layouts.size();
+	m_remainingSets = maxSets * uint32_t(layouts.size());
 
 	NewDescriptorPool(poolSize, m_remainingSets, &m_descPoolHandle);
 }
